worker/requirements.txt

psycopg2-binary
redis
requests
uuid
